api endpoint started to added & some db change and effected layers

+++ Testgram.Core/IRepositories/ICommentRepository.cs
@@ -7,8 +7,8 @@ namespace Testgram.Core.IRepositories
     public interface ICommentRepository : IRepository<Comment>
     {
         Task<IEnumerable<Comment>> GetAllCommentsAsync();
-        Task<IEnumerable<Comment>> GetCommentsByPostIdAsync(int id);
-        Task<IEnumerable<Comment>> GetCommentsByUserIdAsync(int id);
-        Task<IEnumerable<Comment>> GetCommentsByParentIdAsync(int id);
+        Task<IEnumerable<Comment>> GetCommentsByPostIdAsync(long id);
+        Task<IEnumerable<Comment>> GetCommentsByUserIdAsync(long id);
+        Task<IEnumerable<Comment>> GetCommentsByParentIdAsync(long id);
     }
 }
