+++ Testgram.Core/IRepositories/IFollowRepository.cs
@@ -7,7 +7,7 @@ namespace Testgram.Core.IRepositories
     public interface IFollowRepository : IRepository<Follow>
     {
         Task<IEnumerable<Follow>> GetAllFollowsAsync();
-        Task<IEnumerable<Follow>> GetFollowsByFollowerIdAsync(int id);
-        Task<IEnumerable<Follow>> GetFollowsByUserIdAsync(int id);
+        Task<IEnumerable<Follow>> GetFollowsByFollowerIdAsync(long id);
+        Task<IEnumerable<Follow>> GetFollowsByUserIdAsync(long id);
     }
 }
