+++ Testgram.Core/IRepositories/ILikeRepository.cs
@@ -7,7 +7,7 @@ namespace Testgram.Core.IRepositories
     public interface ILikeRepository : IRepository<Likes>
     {
         Task<IEnumerable<Likes>> GetAllLikesAsync();
-        Task<IEnumerable<Likes>> GetLikesByPostIdAsync(int id);
-        Task<IEnumerable<Likes>> GetLikesByUserIdAsync(int id);
+        Task<IEnumerable<Likes>> GetLikesByPostIdAsync(long id);
+        Task<IEnumerable<Likes>> GetLikesByUserIdAsync(long id);
     }
 }
