+++ Testgram.Core/IRepositories/IPostRepository.cs
@@ -8,8 +8,8 @@ namespace Testgram.Core.IRepositories
     public interface IPostRepository : IRepository<Post>
     {
         Task<IEnumerable<Post>> GetAllPostsAsync();
-        Task<Post> GetPostByIdAsync(int id);
-        Task<IEnumerable<Post>> GetPostsByUserIdAsync(int userId);
+        Task<Post> GetPostByIdAsync(long id);
+        Task<IEnumerable<Post>> GetPostsByUserIdAsync(long userId);
         Task<IEnumerable<Post>> GetPostsAfterDateAsync(DateTime dateTime);
     }
 }
