+++ Testgram.Core/IRepositories/IProfileRepository.cs
@@ -7,7 +7,7 @@ namespace Testgram.Core.IRepositories
     public interface IProfileRepository : IRepository<Profile>
     {
         Task<IEnumerable<Profile>> GetAllProfilesAsync();
-        Task<Profile> GetProfileByIdAsync(int id);
+        Task<Profile> GetProfileByIdAsync(long id);
         Task<Profile> GetProfileByUsernameAsync(string username);
         Task<Profile> GetProfileByEmailAsync(string email);
     }
