+++ Testgram.Core/IServices/ICommentService.cs
@@ -7,9 +7,9 @@ namespace Testgram.Core.IServices
     public interface ICommentService
     {
         Task<IEnumerable<Comment>> GetAllComments();
-        Task<IEnumerable<Comment>> GetAllCommentsByPostId(int postId);
-        Task<IEnumerable<Comment>> GetAllCommentsByUserId(int userId);
-        Task<IEnumerable<Comment>> GetAllCommentsByParentComment(int commentId);
+        Task<IEnumerable<Comment>> GetAllCommentsByPostId(long postId);
+        Task<IEnumerable<Comment>> GetAllCommentsByUserId(long userId);
+        Task<IEnumerable<Comment>> GetAllCommentsByParentComment(long commentId);
         Task<Comment> CreateComment(Comment comment);
         Task UpdateComment(Comment commentToBeUpdated, Comment comment);
         Task DeleteComment(Comment comment);
