+++ Testgram.Core/IServices/IFollowService.cs
@@ -7,8 +7,8 @@ namespace Testgram.Core.IServices
     public interface IFollowService
     {
         Task<IEnumerable<Follow>> GetAllFollows();
-        Task<IEnumerable<Follow>> GetFollowByUserId(int userId);
-        Task<IEnumerable<Follow>> GetFollowByFollowerId(int followerId);
+        Task<IEnumerable<Follow>> GetFollowByUserId(long userId);
+        Task<IEnumerable<Follow>> GetFollowByFollowerId(long followerId);
         Task<Follow> CreateFollow(Follow follow);
         Task DeleteFollow(Follow follow);
     }
