+++ Testgram.Core/IServices/IProfileService.cs
@@ -7,7 +7,7 @@ namespace Testgram.Core.IServices
     public interface IProfileService
     {
         Task<IEnumerable<Profile>> GetAllProfiles();
-        Task<Profile> GetProfileById(int id);
+        Task<Profile> GetProfileById(long id);
         Task<Profile> GetProfileByUsername(string username);
         Task<Profile> GetProfileByEmail(string email);
         Task<Profile> CreateProfile(Profile profile);
